refactor(gui): fold Logs + Info page builders into their mixins
Extend LogsMixin with _build_logs_page (Terminal tab — title + quick-
button row + log_output QTextEdit + cmd_input row) and InfoPageMixin
with _build_info_page (Info tab — 6-card scaffold + async refresh
kick-off). Both builders previously lived on WinpodxWindow alongside
the handlers they wire up; pairing builder + behaviour in the same
mixin keeps each page a self-contained unit.

No behaviour change; the v0.1.9.1 QTimer.singleShot(0, _refresh_info)
deferred-fetch ordering is preserved exactly. Tests: 1096 passed.

main_window.py: 411 -> 243 lines (-168). Cumulative: 2745 -> 243 (-91%).
